fix: move resolveStyleNames from saperate file to styles.ts
