+++ packages/framework/src/shared/index.ts
@@ -2,4 +2,3 @@ export * from "./actions";
 export * from "./models";
 export * from "./dto";
 export * from "./common";
-export * from "./styles";
